[mle-router] adding FindNeighbor() to find a neighbor matching a state filter (#4440)
The newly added `FindNeighbor()` method searches within child table,
router table (independent of the current role of device), and also
between parent and parent candidate for a neighbor matching a given
MAC address and a neighbor state filter.

The behavior of this method is different from `GetNeighbor()` which
only accepts neighbors in valid or restoring states and searches
within child and router tables only when device is in router or leader
roles.