[discover-scanner] allow filtering based on a user given ID (#5075)
This commit enhances the filtering behavior of Discover Scan. When
filtering is enabled, MLE Discovery Responses with steering data
(bloom filter) not containing a given ID (MAC Extended Address) are
filtered. This commit updates the `DiscoverScanner::Discover()` to
allow the ID used for filtering to be optionally specified by the
caller. It can be set to NULL (for default behavior) where the hash
of factory-assigned EUI64 of the device would be used.