[mle] change ProcessRouteTlv() to update the router/neighbor entry
This commit changes the `ProcessRouteTlv()` to require caller to pass
in a reference to a `Router` or `Neighbor`.  This method parses a
given `RouterTlv` which can then shift/move the `Router` entries in
the router table (causing pointers to router entries to become
invalid). When the passed in `aRouter` pointer is not null, this
method ensures that it is updated to points to the same router entry
as before.