[logging] update OT logging (use functions)
This commit updates and enhances OT logging by implementing some of
more common operations as functions (instead of macros) and preparing
the entire log line (appending log level and/or log region to log
string if the corresponding OT config options are enabled) before
passing it to platform layer.

The OT logging previously used macro definitions all the way to
platform layer call to `otPlatLog` (which itself could be defined as a
macro through an OT config). This model was adopted to address a
restriction/requirement on certain platforms (e.g., to enable log
tokenization at platform layer). However, using macros adds an
overhead on all other platforms that do not have such a requirement.

 This commit adds `OPENTHREAD_CONFIG_LOG_DEFINE_AS_MACRO_ONLY` which
 when enabled (set to 1) keeps the implementation as before (i.e.,
 macros are used in the logging module). By default this is disabled
 which enables the new implementation added from this commit.

This commit also adds `OPENTHREAD_CONFIG_PLAT_LOG_MACRO_NAME` which
replaces the previous `OPENTHREAD_CONFIG_PLAT_LOG_FUNCTION_NAME`. This
config is used to override the `otPlatLog` and define it as macro. A
build-time check is added in the `config-check.h` to help any project
that is integrating OT and may be using/relying on macro name override
behavior to detect the change from this commit and update its config
definitions.