[net] SRP client implementation
This commit adds support for SRP (Service Registration Protocol)
client in OpenThread. It adds `Srp::Client` class along with its
public OT APIs and definitions under `otSrpClient{Name}`.

The implementation allows user to provide host info (host name and a
list of host IPv6 addresses) along with a list of services to be
registered with an SRP server. Services and/or host addresses can be
added or removed during operation of client.

Public APIs allow user to can get the list of services and host info
and their current state (e.g, is it registered with server, being
added or removed, etc). Also user can provide a callback to get
notified whenever there is change (e.g., a service is successfully
registered with server).

When there is change (e.g., a new service is added/removed) that
requires an update, the SRP client will wait for a short delay before
sending an SRP update to server. This allows the user to provide more
change that are then all sent in same  update message. The
implementation handles retries in case of different errors and
failures. An exponentially increasing retry wait interval model (with
configurable min/max and growth factor) is implemented.

The implementation also manages the lease renew time for each service
and refreshes (re-registers) services with server in new SRP updates
when lease is about to expire. It also supports opportunistic early
refresh mechanism, where when sending an SRP update, the services that
are not yet expired but are close, are allowed to refresh early and
are included in the SRP update. This helps place more services on the
same lease refresh schedule reducing number of messages sent to the
SRP server. This behavior (whether to allow early refresh or not, or
always include all services and its parameters) can be controlled
through a set of OT config definitions.