[dns-header] new helper method to find i-th record matching name and type
This commit adds a new template helper method in `ResoureRecord` to
search in a given message starting at a given offset and up to a
maximum given number of records, for the i-th occurrence of a specific
record (of given type) matching a given record name. If found, it reads
the record from the message into a given record object.

The unit test `test_dns` is updated to cover the behavior of newly
added helper method.