[tmf] add `AnycastLocator` module
This commit adds a new class `AnycastLocatr` which can be used to
locate the closest destination of an anycast IPv6 address (i.e., find
the destination's mesh local EID and RLOC16). The closest destination
is determined based on the the current routing table and path costs
within the Thread mesh.

The implementation uses a CoAP confirmable post request to a newly
added URI path ("a/yl"). The destination IPv6 address of such as
request message is set to the anycast address to be located. The
receiver of the request message sends a CoAP response which includes
the "Mesh Local EID"  and "Thread RLOC16" TLVs.

This commit also adds support this new feature in CLI (adding a new
`locate <anycast-addr>` command).