(WITH INT) - [cli] adding helper `ProcessGet/Set()` to handle simpler commands
This commit adds a set of helper methods in `Cli` to process simple
commands where we are getting and/or setting a single `uint` value.
This replaces common pattern in processing of commands where if there
are no arguments, we output the value we retrieve by calling a related
getter OT API, otherwise we parse the argument (to get the proper
`uint` value) and call a related setter OT API. For example, for
"channel" command, we use the `otLinkGetChannel()` as the getter and
`otLinkSetChannel()` as setter.