[netdata] mechanism to detect & signal when network data gets full
This commit adds a new mechanism to detect when Network Data (local
or leader) gets full and signals by invoking a callback which can be
set with a newly added OT API.

New method is added in `NetworkData::Leader` to determine whether
registering local Network Data into the current Leader data can be
successful or if it may lead to Network Data getting full. This is
determined by device emulating the leader behavior and registering
the local Network Data into a clone of the current leader's Network
Data.

[netdata] new feature to optimize published routes when netdata gets full

This commit adds a new feature in `NetworkData::Publisher` to monitor
when Network Data gets full and optimize published external routes
entries and replace them with a set of compact (shorter length)
prefixes.

`OPENTHREAD_CONFIG_NETDATA_PUBLISHER_OPTIMIZE_ROUTES_ON_FULL_NETDATA`
controls this feature and by default it is enabled when device is
configured as BR (i.e.,  `BORDER_ROUTING`/`RoutingManager` is
enabled).

The compact prefix list is determined using the following algorithm:
- The initial "compact prefix length" is set to 48.
- All route prefixes (excluding `::/0`) are shortened to the "compact
  prefix length" if they are longer.
- All repeating prefixes are removed.
- Any prefix that is contained within another shorter prefix is removed.
- If the new set can not still fit in Network Data, the "compact prefix
  length" is shortened to 10 and then to 7.

Once in optimized mode, `Publisher` will periodically check if it can
again go back to normal mode by trying to publish all routes and
checking if network data gets full. The default period is set to 4
hours (with a jitter of 15 minutes).

This commit adds a test-case in `test_network_data` to validate the
behavior of newly added mechanism.

[netdata] new feature to optimize published routes when netdata gets full

This commit adds a new feature in `NetworkData::Publisher` to monitor
when Network Data gets full and optimize published external routes
entries and replace them with a set of compact (shorter length)
prefixes.

`OPENTHREAD_CONFIG_NETDATA_PUBLISHER_OPTIMIZE_ROUTES_ON_FULL_NETDATA`
controls this feature and by default it is enabled when device is
configured as BR (i.e.,  `BORDER_ROUTING`/`RoutingManager` is
enabled).

The compact prefix list is determined using the following algorithm:
- The initial "compact prefix length" is set to 48.
- All route prefixes (excluding `::/0`) are shortened to the "compact
  prefix length" if they are longer.
- All repeating prefixes are removed.
- Any prefix that is contained within another shorter prefix is removed.
- If the new set can not still fit in Network Data, the "compact prefix
  length" is shortened to 10 and then to 7.

Once in optimized mode, `Publisher` will periodically check if it can
again go back to normal mode by trying to publish all routes and
checking if network data gets full. The default period is set to 4
hours (with a jitter of 15 minutes).

This commit adds a test-case in `test_network_data` to validate the
behavior of newly added mechanism.