[ip6] update receiving of tunnel message in `HandleDatagram()`
This commit updates how `Ip6::HandleDatagram()` processes a tunneled
IP-in-IP message. It is possible that such a message is marked for
both `receive` and `forwardThread`, e.g., a multicast message with
host origin and multicast loop flag.  This commit ensures that such a
message is handled properly. Without the change in the commit, the message
would be received but not forwarded to Thread mesh. The multicast behavior
continues to work fine due to MPL retransmission which would be then
forwarded to the thread mesh but the first message would not be.

This commit changes the code, so that if we need to `forwardThread` we
create a copy of the message by cloning it, otherwise we take ownership
of the original message and and use it directly. The encapsulating
header is then removed before processing the embedded message by
calling the `HandleDatagram()` recursively.