[netif] simplify applying mesh-local prefix change to addresses
This commit simplifies how the `Netif` unicast and multicast addresses
that use the mesh-local prefix are updated when the mesh-local prefix
is changed.

The `Netif::UnicastAddress` class now includes a `mMeshLocal` flag,
which is set for all mesh-local addresses (RLOCs, ALOCs, and ML-EID).
When the mesh-local prefix is changed, `Mle` will inform
`ThreadNetif` to apply the new prefix. `ThreadNetif` will then update
all of the assigned unicast addresses that are marked as mesh-local,
as well as all of the subscribed mesh-local prefix-based multicast
addresses (such as link-local or realm-local All Thread Nodes
addresses). The `Netif` class first signals the removal of the
previous address based on the old prefix, and then the addition of
the new address with the new mesh-local prefix.

This change simplifies the code, in particular the modules such as
`Commissioner`, `BorderAgent` and `BackboneRouter::Local`, which need
to manage specific ALOC addresses. These modules no longer need to be
informed when the mesh-local prefix is changed to update their
address, as it will be done by the `Netif` class directly.