[encoding] avoid the use of `using` for `Big/LittleEndian` functions
This commit update the code to remove the use of `using` for
`BigEndian` or `LittleEndian` functions such as `HostSwap` or
`ReadUint`. Since these functions often used in header files,
using direct namespace, makes the code safer in case incorrect
`using` is adopted based on order of the incldued headers.
In general, avoiding `using` keyword in header is recommended.

This commit also removes the `Encoding` namespace to make the
full name of method shorter.